:lipstick: thinner investments table
