+++ app/[usernameSlug]/user-active-bids.tsx
@@ -1,9 +1,11 @@
 'use client'
-import { BidAndProject, deleteBid } from '@/db/bid'
-import { BidText, TableRow } from '@/components/tables'
-import { RoundTag } from '@/components/tags'
+import { Bid, BidAndProject, deleteBid } from '@/db/bid'
+import { ThickTableRow } from '@/components/tables'
+import { RoundTag, Tag } from '@/components/tags'
 import { useSupabase } from '@/db/supabase-provider'
 import { useRouter } from 'next/navigation'
+import { formatMoney } from '@/utils/formatting'
+import { Project } from '@/db/project'
 
 export function ActiveBids(props: {
   bids: BidAndProject[]
@@ -13,7 +15,7 @@ export function ActiveBids(props: {
   const { supabase } = useSupabase()
   const router = useRouter()
   const bidsDisplay = bids.map((bid) => (
-    <TableRow
+    <ThickTableRow
       key={bid.id}
       title={bid.projects.title}
       subtitle={
@@ -43,4 +45,59 @@ export function ActiveBids(props: {
       </table>
     </div>
   )
+}
+
+export function BidText(props: {
+  bid: Bid
+  stage: string
+  projectType: Project['type']
+  showValuation: boolean
+}) {
+  const { bid, stage, projectType, showValuation } = props
+  switch (stage) {
+    case 'proposal':
+      if (projectType === 'grant') {
+        return (
+          <div className="flex items-center">
+            <p className="text-sm text-gray-500">
+              Offered&nbsp;
+              <span className="text-black">{formatMoney(bid.amount)}</span>
+            </p>
+          </div>
+        )
+      }
+      return (
+        <div className="flex items-center">
+          <p className="text-sm text-gray-500">
+            Bid&nbsp;
+            <span className="text-black">{formatMoney(bid.amount)}</span>
+            {showValuation && (
+              <span>
+                &nbsp;@&nbsp;
+                <span className="text-black">{formatMoney(bid.valuation)}</span>
+                &nbsp;valuation
+              </span>
+            )}
+          </p>
+        </div>
+      )
+    case 'active':
+      return (
+        <div className="flex items-center text-sm text-gray-500">
+          Offer to&nbsp;
+          <Tag
+            text={bid.type === 'buy' ? 'BUY' : 'SELL'}
+            color={bid.type === 'buy' ? 'emerald' : 'rose'}
+          />
+          <span className="text-black">&nbsp;{formatMoney(bid.amount)}</span>
+          <span>
+            &nbsp;@&nbsp;
+            <span className="text-black">{formatMoney(bid.valuation)}</span>
+            &nbsp;valuation
+          </span>
+        </div>
+      )
+    default:
+      return <></>
+  }
 }
